feat : refactor trba_detector / triton_client
